+++ src/main/java/com/photosend/photosendserver01/domains/catalog/presentation/request_response/ProductSummaryForCatalog.java
@@ -12,7 +12,7 @@ public class ProductSummaryForCatalog {
     private long productId;
     private String name;
     private String brand;
-    private int price;
+    private Double price;
     private String productImagePath;
     private LocalDateTime uploadedTime;
 }
